+++ src/main/java/com/ibm/devops/connect/DevOpsGlobalConfiguration.java
@@ -34,6 +34,9 @@ import com.cloudbees.plugins.credentials.common.StandardUsernamePasswordCredenti
 import com.cloudbees.plugins.credentials.domains.URIRequirementBuilder;
 import hudson.security.ACL;
 import jenkins.model.Jenkins;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
 
 /**
  * Created by lix on 7/20/17.
@@ -49,44 +52,17 @@ public class DevOpsGlobalConfiguration extends GlobalConfiguration {
     private String rabbitMQPort;
     private String rabbitMQHost;
     private String apiToken;
+    private List<Entry> entries= new ArrayList<>();
 
     public DevOpsGlobalConfiguration() {
         load();
     }
-
-    public String getSyncId() {
-    	return syncId;
+    public List<Entry> getEntries() {
+    	return entries;
     }
 
-    public void setSyncId(String syncId) {
-        this.syncId = syncId;
-        save();
-    }
-
-    public String getSyncToken() {
-    	return syncToken;
-    }
-
-    public void setSyncToken(String syncToken) {
-        this.syncToken = syncToken;
-        save();
-    }
-
-    public String getApiToken() {
-        return apiToken;
-    }
-
-    public void setApiToken(String apiToken) {
-        this.apiToken = apiToken;
-        save();
-    }
-
-    public String getBaseUrl() {
-    	return baseUrl;
-    }
-
-    public void setBaseUrl(String baseUrl) {
-        this.baseUrl = baseUrl;
+    public void setEntries(List<Entry> entries) {
+        this.entries = entries;
         save();
     }
 
@@ -99,69 +75,33 @@ public class DevOpsGlobalConfiguration extends GlobalConfiguration {
         save();
     }
 
-    public String getRabbitMQPort() {
-        return rabbitMQPort;
-    }
-
-    public String getRabbitMQHost() {
-        return rabbitMQHost;
-    }
-
-    public void setRabbitMQPort(String rabbitMQPort) {
-        this.rabbitMQPort = rabbitMQPort;
-        save();
-    }
-
-    public void setRabbitMQHost(String rabbitMQHost) {
-        this.rabbitMQHost = rabbitMQHost;
-        save();
-    }
 
     @Override
     public boolean configure(StaplerRequest req, JSONObject formData) throws FormException {
         // To persist global configuration information,
         // set that to properties and call save().
-        syncId = formData.getString("syncId");
+        /*syncId = formData.getString("syncId");
         syncToken = formData.getString("syncToken");
         baseUrl = formData.getString("baseUrl");
         credentialsId = formData.getString("credentialsId");
         rabbitMQPort = formData.getString("rabbitMQPort");
         rabbitMQHost = formData.getString("rabbitMQHost");
         apiToken = formData.getString("apiToken");
+        //entries = formData.getString("entries");*/
+        setEntries(Collections.<Entry>emptyList());
+        req.bindJSON(this, formData);
         save();
 
         reconnectCloudSocket();
 
-        return super.configure(req,formData);
+        //return super.configure(req,formData);
+        return true;
     }
 
     // for the future multi-region use
     public ListBoxModel doFillRegionItems() {
         ListBoxModel items = new ListBoxModel();
         return items;
-    }
-
-    public FormValidation doTestConnection(@QueryParameter("syncId") final String syncId,
-        @QueryParameter("syncToken") final String syncToken,
-        @QueryParameter("baseUrl") final String baseUrl)
-    throws FormException {
-        try {
-            boolean connected = CloudPublisher.testConnection(syncId, syncToken, baseUrl);
-            if (connected) {
-                boolean amqpConnected = CloudSocketComponent.isAMQPConnected();
-
-                String rabbitMessage = "Not connected to RabbitMQ. Unable to run Jenkins jobs from UCV.";
-                if(amqpConnected) {
-                    rabbitMessage = "Connected to RabbitMQ successfully. Ready to run Jenkins jobs from UCV.";
-                }
-
-                return FormValidation.ok("Successful connection to Velocity Services.\n" + rabbitMessage);
-            } else {
-                return FormValidation.error("Could not connect to Velocity.  Please check your URL and credentials provided.");
-            }
-        } catch (Exception e) {
-            return FormValidation.error("Could not connect to Velocity : " + e.getMessage());
-        }
     }
 
     /**
@@ -199,11 +139,4 @@ public class DevOpsGlobalConfiguration extends GlobalConfiguration {
         connectComputerListener.onOnline(Jenkins.getInstance().toComputer());
     }
 
-    public boolean isConfigured() {
-        return StringUtils.isNotEmpty(this.syncId) &&
-               StringUtils.isNotEmpty(this.syncToken) &&
-               StringUtils.isNotEmpty(this.baseUrl) &&
-               StringUtils.isNotEmpty(this.credentialsId) &&
-               StringUtils.isNotEmpty(this.apiToken);
-    }
 }
