+++ src/main/resources/com/ibm/devops/connect/Entry/config.jelly
@@ -1,0 +1,33 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
+    <f:entry title="Connect to UrbanCode Velocity Instance">
+        <f:entry title="Integration ID" field="syncId">
+            <f:textbox />
+        </f:entry>
+        <f:entry title="Integration Token" field="syncToken">
+            <f:password />
+        </f:entry>
+        <f:entry title="Velocity Base URL" field="baseUrl">
+            <f:textbox />
+        </f:entry>
+        <f:entry title="${%User Access Key}" field="apiToken">
+            <f:password />
+        </f:entry>
+        <f:advanced>
+            <f:entry title="${%Rabbit MQ Port}" field="rabbitMQPort">
+                <f:textbox />
+            </f:entry>
+            <f:entry title="${%Rabbit MQ Host}" field="rabbitMQHost">
+                <f:textbox />
+            </f:entry>
+        </f:advanced>
+        <f:entry title="">
+            <div style="text-align: right">
+            <f:validateButton
+                title="${%Test Connection}" progress="${%Testing...}"
+                method="testConnections" with="syncId,syncToken,baseUrl,rabbitMQHost,rabbitMQPort"/>
+            </div>
+        </f:entry>
+    </f:entry>
+       
+</j:jelly>
